Set up search repository
